+++ src/components/QueuePanel.tsx
@@ -1,7 +1,6 @@
-import { getPlaylists, getPlaylist, updatePlaylist, deletePlaylist } from '../api/subsonicPlaylists';
+import { getPlaylist, updatePlaylist } from '../api/subsonicPlaylists';
 import { buildCoverArtUrl, coverArtCacheKey } from '../api/subsonicStreamUrl';
 import { getAlbum } from '../api/subsonicLibrary';
-import type { SubsonicPlaylist } from '../api/subsonicTypes';
 import { registerQueueListScrollTopReader, consumePendingQueueListScrollTop } from '../store/queueUndo';
 import { songToTrack } from '../utils/songToTrack';
 import type { Track } from '../store/playerStoreTypes';
@@ -12,7 +11,7 @@ import { useOrbitStore } from '../store/orbitStore';
 import OrbitGuestQueue from './OrbitGuestQueue';
 import OrbitQueueHead from './OrbitQueueHead';
 import HostApprovalQueue from './HostApprovalQueue';
-import { Play, Music, Star, X, Trash2, Save, FolderOpen, Shuffle, Infinity, Waves, MicVocal, ListMusic, Check, ListPlus, MoveRight, Radio, HardDrive, ChevronDown, Info, Share2 } from 'lucide-react';
+import { Play, Music, Trash2, Save, FolderOpen, Shuffle, Infinity, Waves, MicVocal, ListMusic, Check, MoveRight, Radio, HardDrive, ChevronDown, Info, Share2 } from 'lucide-react';
 import { usePlaylistStore } from '../store/playlistStore';
 import { useCachedUrl } from './CachedImage';
 import { useTranslation } from 'react-i18next';
@@ -32,165 +31,14 @@ import { useLuckyMixStore } from '../store/luckyMixStore';
 import { useQueueToolbarStore, QueueToolbarButtonId } from '../store/queueToolbarStore';
 import { loudnessGainPlaceholderUntilCacheDb } from '../utils/loudnessPlaceholder';
 import { effectiveLoudnessPreAnalysisAttenuationDb } from '../utils/loudnessPreAnalysisSlider';
-
-function formatTime(seconds: number): string {
-  if (!seconds || isNaN(seconds)) return '0:00';
-  const m = Math.floor(seconds / 60);
-  const s = Math.floor(seconds % 60);
-  return `${m}:${s.toString().padStart(2, '0')}`;
-}
-
-function formatQueueReplayGainParts(track: Track, t: TFunction): string[] {
-  const parts: string[] = [];
-  const fmtDb = (db: number) => `${db >= 0 ? '+' : ''}${db.toFixed(1)}`;
-  if (track.replayGainTrackDb != null) {
-    parts.push(t('queue.rgTrack', { db: fmtDb(track.replayGainTrackDb) }));
-  }
-  if (track.replayGainAlbumDb != null) {
-    parts.push(t('queue.rgAlbum', { db: fmtDb(track.replayGainAlbumDb) }));
-  }
-  if (track.replayGainPeak != null) {
-    parts.push(t('queue.rgPeak', { pk: track.replayGainPeak.toFixed(3) }));
-  }
-  return parts;
-}
-
-function renderStars(rating?: number) {
-  if (!rating) return null;
-  const stars = [];
-  for (let i = 1; i <= 5; i++) {
-    stars.push(
-      <Star 
-        key={i} 
-        size={12} 
-        fill={i <= rating ? 'var(--ctp-yellow)' : 'none'} 
-        color={i <= rating ? 'var(--ctp-yellow)' : 'var(--text-muted)'} 
-      />
-    );
-  }
-  return <div style={{ display: 'flex', gap: '2px', alignItems: 'center' }}>{stars}</div>;
-}
-
-function SavePlaylistModal({ onClose, onSave }: { onClose: () => void, onSave: (name: string) => void }) {
-  const { t } = useTranslation();
-  const [name, setName] = useState('');
-  return (
-    <div className="modal-overlay" onClick={onClose} role="dialog" aria-modal="true">
-      <div className="modal-content" onClick={e => e.stopPropagation()} style={{ maxWidth: '400px' }}>
-        <button className="modal-close" onClick={onClose}><X size={18} /></button>
-        <h3 style={{ marginBottom: '1rem', fontFamily: 'var(--font-display)' }}>{t('queue.savePlaylist')}</h3>
-        <input 
-          type="text" 
-          className="live-search-field" 
-          placeholder={t('queue.playlistName')} 
-          value={name} 
-          onChange={e => setName(e.target.value)}
-          autoFocus
-          onKeyDown={e => e.key === 'Enter' && name.trim() && onSave(name.trim())}
-          style={{ width: '100%', marginBottom: '1rem', padding: '10px 16px' }}
-        />
-        <div style={{ display: 'flex', justifyContent: 'flex-end', gap: '8px' }}>
-          <button className="btn btn-ghost" onClick={onClose}>{t('queue.cancel')}</button>
-          <button className="btn btn-primary" onClick={() => name.trim() && onSave(name.trim())}>{t('queue.save')}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function LoadPlaylistModal({ onClose, onLoad }: { onClose: () => void, onLoad: (id: string, name: string, mode: 'replace' | 'append') => void }) {
-  const { t } = useTranslation();
-  const [playlists, setPlaylists] = useState<SubsonicPlaylist[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [filter, setFilter] = useState('');
-  const [confirmDelete, setConfirmDelete] = useState<{ id: string; name: string } | null>(null);
-
-  const fetchPlaylists = () => {
-    setLoading(true);
-    getPlaylists().then(data => {
-      setPlaylists(data);
-      setLoading(false);
-    }).catch(e => {
-      console.error(e);
-      setLoading(false);
-    });
-  };
-
-  useEffect(() => {
-    fetchPlaylists();
-  }, []);
-
-  const handleDelete = async (id: string, name: string) => {
-    setConfirmDelete({ id, name });
-  };
-
-  const confirmDeletePlaylist = async () => {
-    if (!confirmDelete) return;
-    await deletePlaylist(confirmDelete.id);
-    setConfirmDelete(null);
-    fetchPlaylists();
-  };
-
-  return (
-    <>
-    <div className="modal-overlay" onClick={onClose} role="dialog" aria-modal="true">
-      <div className="modal-content" onClick={e => e.stopPropagation()} style={{ maxWidth: '560px', width: '90vw' }}>
-        <button className="modal-close" onClick={onClose}><X size={18} /></button>
-        <h3 style={{ marginBottom: '1rem', fontFamily: 'var(--font-display)' }}>{t('queue.loadPlaylist')}</h3>
-        {!loading && playlists.length > 0 && (
-          <input
-            type="text"
-            className="live-search-field"
-            placeholder={t('queue.filterPlaylists')}
-            value={filter}
-            onChange={e => setFilter(e.target.value)}
-            autoFocus
-            style={{ width: '100%', marginBottom: '0.75rem', padding: '8px 14px' }}
-          />
-        )}
-        {loading ? (
-          <p style={{ color: 'var(--text-muted)' }}>{t('queue.loading')}</p>
-        ) : playlists.length === 0 ? (
-          <p style={{ color: 'var(--text-muted)' }}>{t('queue.noPlaylists')}</p>
-        ) : (
-          <div style={{ display: 'flex', flexDirection: 'column', gap: '8px', maxHeight: '300px', overflowY: 'auto' }}>
-            {playlists.filter(p => p.name.toLowerCase().includes(filter.toLowerCase())).map(p => (
-              <div key={p.id} style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', padding: '8px 12px', background: 'var(--ctp-surface1)', borderRadius: 'var(--radius-md)' }}>
-                <span style={{ fontWeight: 500 }} className="truncate" data-tooltip={p.name}>{p.name}</span>
-                <div style={{ display: 'flex', gap: '4px', flexShrink: 0 }}>
-                  <button className="nav-btn" onClick={() => onLoad(p.id, p.name, 'replace')} data-tooltip={t('queue.load')} style={{ width: '28px', height: '28px', background: 'transparent' }}><Play size={14} /></button>
-                  <button className="nav-btn" onClick={() => onLoad(p.id, p.name, 'append')} data-tooltip={t('queue.appendToQueue')} style={{ width: '28px', height: '28px', background: 'transparent' }}><ListPlus size={14} /></button>
-                  <button className="nav-btn" onClick={() => handleDelete(p.id, p.name)} data-tooltip={t('queue.delete')} style={{ width: '28px', height: '28px', background: 'transparent', color: 'var(--ctp-red)' }}><Trash2 size={14} /></button>
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-
-    {confirmDelete && (
-      <div className="modal-overlay" onClick={() => setConfirmDelete(null)} role="dialog" aria-modal="true">
-        <div className="modal-content" onClick={e => e.stopPropagation()} style={{ maxWidth: '360px' }}>
-          <button className="modal-close" onClick={() => setConfirmDelete(null)}><X size={18} /></button>
-          <h3 style={{ marginBottom: '0.5rem', fontFamily: 'var(--font-display)' }}>{t('queue.delete')}</h3>
-          <p style={{ color: 'var(--text-secondary)', marginBottom: '1.25rem', lineHeight: 1.5 }}>
-            {t('queue.deleteConfirm', { name: confirmDelete.name })}
-          </p>
-          <div style={{ display: 'flex', justifyContent: 'flex-end', gap: '8px' }}>
-            <button className="btn btn-ghost" onClick={() => setConfirmDelete(null)}>{t('queue.cancel')}</button>
-            <button className="btn btn-primary" style={{ background: 'var(--danger)', borderColor: 'var(--danger)' }} onClick={confirmDeletePlaylist}>
-              {t('queue.delete')}
-            </button>
-          </div>
-        </div>
-      </div>
-    )}
-    </>
-  );
-}
-
-type DurationMode = 'total' | 'remaining' | 'eta';
+import {
+  DurationMode,
+  formatTime,
+  formatQueueReplayGainParts,
+  renderStars,
+} from '../utils/queuePanelHelpers';
+import { SavePlaylistModal } from './queuePanel/SavePlaylistModal';
+import { LoadPlaylistModal } from './queuePanel/LoadPlaylistModal';
 
 interface QueueHeaderProps {
   queue: Track[];
